fix: CreateAccessToken is more generic for other use cases (#79)
* fix: CreateAccessToken is more generic for other use cases

* fix: ScopesToPermissions now errors out if no permissions are valid

returning an error early means that we don't attempt to vend a token with no permissions, preventing potential silent errors during token vending or use

* fix: GetRepoNames succeeds if at least one URL is valid

this prevents a single typo interferring with the entire token generation process. additionally, we now validate if the url is a valid github url, previously this wasn't the case

* fix: ScopesToPermissions now accounts for snake case

previously multi-word permissions such as 'pull_request' wouldn't be captialised correctly. converting to pascal case solves for this

* chore: remove link from log

* test: confirm that token is empty string when request fails

* chore: shorten structured logs

* fix: use slices.Contains over bespoke function

not sure how that escaped me initially